+++ decor/applications/lobsters/microbenchmarks/queriers/vote.rs
@@ -5,9 +5,9 @@ use mysql::prelude::*;
 use std::*;
 //use log::{warn, debug};
 
-pub fn vote_on_story(db: &mut mysql::Conn, acting_as: Option<u32>, story_id: u32, pos: bool) -> Result<(), mysql::Error> {
+pub fn vote_on_story(db: &mut mysql::Conn, acting_as: Option<u64>, story_id: u64, pos: bool) -> Result<(), mysql::Error> {
     let user = acting_as.unwrap();
-    let (author, score, story) : (u32, f64, u32) = db.query(format!(
+    let (author, score, story) : (u64, f64, u64) = db.query(format!(
                 "SELECT `stories`.user_id, stories.hotness, stories.id \
                  FROM `stories` \
                  WHERE `stories`.`short_id` = {}",
